+++ src/env.js
@@ -32,6 +32,20 @@ export default {
         apiKey: process.env.APM_API_KEY,
         secretToken: process.env.APM_SECRET_TOKEN,
     },
+    email:{
+        type: process.env.EMAIL_TYPE || "OAuth2",
+        host: process.env.EMAIL_HOST || "smtp.gmail.com",
+        port: process.env.EMAIL_PORT ||  465,
+        secure: process.env.EMAIL_SECURE === "true",
+        notificator: process.env.EMAIL_USER,
+        pass: process.env.EMAIL_PASSWORD,
+        OAuth2:{
+          clientId:  process.env.EMAIL_OAUTH2_CLIENTID,
+          clientSecret: process.env.EMAIL_OAUTH2_CLIENTSECRET,
+          refreshToken: process.env.EMAIL_OAUTH2_REFRESHTOKEN,
+          redirectUri: process.env.EMAIL_OAUTH2_REDIRECT_URI || "https://developers.google.com/oauthplayground"
+        },
+    },
     system:{
         files:{
             default: url+"/static/uploads/system/default.png",