+++ src/job/index.js
@@ -5,6 +5,7 @@ import * as updateInvestment from "./jobs/updateInvestment";
 import * as updateDividends from "./jobs/updateDividends";
 import * as asyncDividendsFiis from "./jobs/asyncDividendsFiis";
 import * as asyncDividendsAcao from "./jobs/asyncDividendsAcao";
+import * as backupData from "./jobs/backupData";
 
 logger.info("Registered service JOB batch is ON");
 
@@ -20,6 +21,9 @@ const jobs = [
     }, null, true, env.timezone),
     new CronJob(asyncDividendsAcao.schedule, async ()=>{
         await asyncDividendsAcao.command();
+    }, null, true, env.timezone),
+    new CronJob(backupData.schedule, async ()=>{
+        await backupData.command();
     }, null, true, env.timezone),
 ];